feat(v1.17.0): 十轮穷举交叉验证 — 81项P0-P3修复
10轮全量审查覆盖：
- P0 25项: Entity注解/运行时崩溃/事务缺失/登录认证/Docker/支付回调/讨论区
- P1 29项: 内存泄漏/信息泄漏/IDOR/N+1/数据覆盖/HSTS/CSP
- P2 13项: CSS变量/nginx/购物车/匿名标志/toast防抖/表单锁定
- P3 14项: serialVersionUID/校验注解/死代码/API对齐

交叉验证通过(R1-R4): 代码质量/DB迁移/安全/跨域一致性

Closes: #v1.17.0

+++ micro-course-api/src/main/java/com/microcourse/dto/CourseReviewCreateRequest.java
@@ -18,7 +18,7 @@ public class CourseReviewCreateRequest {
     @NotNull(message = "评分不能为空")
     @Min(value = 1, message = "评分最小为1")
     @Max(value = 5, message = "评分最大为5")
-    private Short rating;
+    private Integer rating;
 
     private String content;
 
@@ -28,8 +28,8 @@ public class CourseReviewCreateRequest {
 
     public Long getCourseId() { return courseId; }
     public void setCourseId(Long courseId) { this.courseId = courseId; }
-    public Short getRating() { return rating; }
-    public void setRating(Short rating) { this.rating = rating; }
+    public Integer getRating() { return rating; }
+    public void setRating(Integer rating) { this.rating = rating; }
     public String getContent() { return content; }
     public void setContent(String content) { this.content = content; }
     public Boolean getIsAnonymous() { return isAnonymous; }
