+++ micro-course-api/src/main/java/com/microcourse/dto/CourseReviewRequest.java
@@ -15,7 +15,7 @@ public class CourseReviewRequest {
     @NotNull(message = "评分不能为空")
     @Min(value = 1, message = "评分最小为1")
     @Max(value = 5, message = "评分最大为5")
-    private Short rating;
+    private Integer rating;
 
     private String content;
 
@@ -23,8 +23,8 @@ public class CourseReviewRequest {
 
     public CourseReviewRequest() {}
 
-    public Short getRating() { return rating; }
-    public void setRating(Short rating) { this.rating = rating; }
+    public Integer getRating() { return rating; }
+    public void setRating(Integer rating) { this.rating = rating; }
     public String getContent() { return content; }
     public void setContent(String content) { this.content = content; }
     public Boolean getIsAnonymous() { return isAnonymous; }
